Share a unittest method via packing to fix unit test dependencies.
